Move codebase into backend package
